Flesh out HubSpot module.
