+++ encrypt/jest.config.js
@@ -1,5 +1,4 @@
 module.exports = async () => {
-    console.log('ENCRYPT JEST');
     return {
         preset: '@friggframework/test-environment',
     };
