+++ frigg-module-hub-spot/credential.test.js
@@ -1,0 +1,23 @@
+const { BaseModelObjectTest } = require('@friggframework/test-environment');
+
+describe('HubSpot Credential', () => {
+    it.skip('does the test', () => {
+    // const testObject = new HubSpotCredentialsTest(
+    //     {
+    //         portal_id: '16854218',
+    //         portal_name: 'Miguel97',
+
+    //         // HS Access Details
+    //         access_token: 'jwt',
+    //         refresh_token: 'jwt referesh',
+    //         expires_in: '2020-07-12',
+    //         expires_at: Date.now(),
+    //         auth_created_at: Date.now(),
+    //         authorized: true,
+    //         auth_is_valid: true,
+    //         user_name: 'Miguel',
+    //     },
+    // );
+    // testObject.test();
+    })
+})
