+++ frigg-module-hub-spot/jest.config.js
@@ -1,3 +1,13 @@
 module.exports = async () => {
-    preset: '@friggframework/test-environment'
-}
+    return {
+        preset: '@friggframework/test-environment',
+        coverageThreshold: {
+            global: {
+                statements: 13,
+                branches: 0,
+                functions: 1,
+                lines: 13,
+            },
+        },
+    };
+};
